Callback mechanism, precision and performance tuning
Provide a way to inject a callback function to `eig` in order to capture
intermediate results.

Implement `eig_reflect__` with unrolled matrix multiplications to see if that is
the bottleneck.

In `eig_2`, pick the root numerically closest to identity in the complex-valued
case and ensure the roots are strictly complex conjugates. This code most
likely suffers from my lack of knowledge of floating point arithmetics.